Extract type alias helper
